Add an ordered TaskSet for coordinating producer and consumers
A TaskSet is a Channel of tasks (little t) that we want to be done
concurrently. The workflow is to have a single `produce!` call
that will continue putting tasks in the queue and waiting if needed
for synchronization checkpoitns. To start consumers, we call `consume!`
and then `take!(::TaskSet)`, which returns a `done` function and our `task`
that we shoudld work on. Once we're done with the task, we call `done()`
to synchronize that our job is done.